pyproject.toml
[project]
name = "pcb-dataset"
version = "0.1.0"
description = "Production-ready pipeline for generating ML datasets of PCB images with segmentation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "h5py>=3.8.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "blenderproc>=2.5.0",
    "noise>=1.2.2",  # Perlin noise generation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
]

[project.scripts]
pcb-gen-single = "pcb_dataset.cli:generate_single"
pcb-gen-batch = "pcb_dataset.cli:generate_batch"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pcb_dataset"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=pcb_dataset",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | pcb2blender
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "pcb2blender",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually
exclude = [
    "pcb2blender/",
    "venv/",
    ".venv/",
]
